% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anscombe.R
\name{anscombe}
\alias{anscombe}
\title{Anscombe's Test for Heteroskedasticity in a Linear Regression Model}
\usage{
anscombe(mainlm, studentise = TRUE)
}
\arguments{
\item{mainlm}{Either an object of \code{\link[base]{class}} "lm", or a list
of two components: a response vector and a design matrix (in that order).
If the latter, the design matrix must begin with a column of 1s if an
intercept is to be included in the linear model. Passing an object of
class "lm" is recommended in applications; passing a list containing
the data is recommended where optimising computational performance is
important.}

\item{studentise}{A logical. Should studentising modification of
\insertCite{Bickel78;textual}{skedastic} be implemented? Defaults to
\code{TRUE}; if \code{FALSE}, the original form of the test proposed by
\insertCite{Anscombe61;textual}{skedastic} is used.}
}
\value{
An object of \code{\link[base]{class}} "htest". If object is not
   assigned, its attributes are displayed in the console as a
   \code{\link[tibble]{tibble}} using \code{\link[broom]{tidy}}.
}
\description{
This function implements the method of
   \insertCite{Anscombe61;textual}{skedastic} for testing for
   heteroskedasticity in a linear regression model, with or without the
   studentising modification of
   \insertCite{Bickel78;textual}{skedastic}.
}
\details{
Anscombe's Test is among the earliest suggestions for heteroskedasticity
   diagnostics in the linear regression model. The test is not based on
   formally derived theory but on a test statistic that Anscombe intuited
   to be approximately standard normal under the null hypothesis of
   homoskedasticity. \insertCite{Bickel78;textual}{skedastic} discusses
   the test and suggests a studentising modification (included in this
   function) as well as a robustifying modification
   (included in \code{\link{bickel}}). The test is two-tailed.

The original form of the test is as follows.
}
\examples{
mtcars_lm <- lm(mpg ~ wt + qsec + am, data = mtcars)
anscombe(mtcars_lm)
anscombe(mtcars_lm, studentise = FALSE)

}
\references{
{\insertAllCited{}}
}
\seealso{
\code{\link{bickel}}, which is a robust extension of this test.
}
