% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blus.R
\name{blus}
\alias{blus}
\title{Compute Best Linear Unbiased Scalar-Covariance (BLUS) residuals from a linear model}
\usage{
blus(mainlm, omit = c("first", "last", "random"), keepNA = TRUE,
  exhaust = NULL)
}
\arguments{
\item{mainlm}{Either an object of \code{\link[base]{class}} "lm", or a list
of two components: a response vector and a design matrix (in that order).
If the latter, the design matrix must begin with a column of 1s if an
intercept is to be included in the linear model. Passing an object of
class "lm" is recommended in applications; passing a list containing
the data is recommended where optimising computational performance is
important.}

\item{omit}{A numeric vector of length \eqn{p} (the number of columns in the
linear model design matrix) giving the indices of \eqn{p} observations to omit in
the BLUS residual vector; or a character partially matching \code{"first"}
(for the first \eqn{p}) observations, \code{"last"} (for the last \eqn{p}
observations), or \code{"random"} (for a random sample of \eqn{p} indices
between 1 and \eqn{n}). Defaults to \code{"first"}.}

\item{keepNA}{A logical. Should BLUS residuals for omitted observations be
returned as \code{NA_real_} to preserve the length of the residual vector?}

\item{exhaust}{An integer. If singular matrices are encountered
using the passed value of \code{omit}, how many random combinations
of \eqn{p} indices should be attempted before an error is thrown? If
\code{NULL} (the default), all possible combinations are attempted
(which could result in very slow execution if \eqn{n} is large).}
}
\value{
A double vector of length \eqn{n} containing the BLUS residuals
   (with \code{NA_real_}) for omitted observations), or a double vector
   of length \eqn{n-p} containing the BLUS residuals only (if \code{keepNA}
   is set to \code{FALSE})
}
\description{
This function computes the Best Linear Unbiased Scalar-Covariance (BLUS)
   residuals from a linear model, as defined in
   \insertCite{Theil65;textual}{skedastic} and explained further in
   \insertCite{Theil68;textual}{skedastic}.
}
\details{
Under the ideal linear model conditions, the BLUS residuals have a scalar
   covariance matrix \eqn{\sigma^2 I} (meaning they have a constant variance
   and are mutually uncorrelated), unlike the OLS residuals, which have
   covariance matrix \eqn{\sigma^2 M} where \eqn{M} is a function of the
   design matrix. Use of BLUS residuals could improve the performance of
   tests for heteroskedasticity and/or autocorrelation in the linear model.
   A linear model with \eqn{n} observations and an \eqn{n\times p} design
   matrix yields only \eqn{n-p} BLUS residuals. The choice of which \eqn{p}
   observations will not be represented in the BLUS residuals is specified
   within the algorithm.
}
\examples{
n <- 20
p <- 4
set.seed(9586)
X <- matrix(data = runif(n * (p - 1)), nrow = n, ncol = p - 1)
# Response values generated under homoskedasticity
y <- rnorm(n, mean = 1 + rowSums(X), sd = 1)
e_tilde <- blus(lm(y ~ X))
}
\references{
{\insertAllCited{}}
}
\seealso{
H. D. Vinod's online article,
   \href{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2412740}{Theil's
   BLUS Residuals and R Tools for Testing and Removing Autocorrelation and
   Heteroscedasticity}, for an alternative function for extracting BLUS
   residuals.
}
