% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{dpeakdist}
\alias{dpeakdist}
\title{Probability mass function of number of peaks in a random series}
\usage{
dpeakdist(n, k)
}
\arguments{
\item{n}{A positive integer representing the number of observations in the
series.}

\item{k}{An integer or a sequence of integers strictly incrementing by 1,
with all values between 0 and \code{n - 1} inclusive. Represents the number
of peaks in the series.}
}
\value{
A double between 0 and 1 representing the probability of exactly
k peaks occurring in a series of \eqn{n} independent and identically
distributed continuous random variables. Computation time is very slow for
\eqn{n > 170}.
}
\description{
This function computes \eqn{P(n,k)} as defined by
\insertCite{Goldfeld65;textual}{skedastic}, i.e. the probability that a
series of \eqn{n} independent and identically distributed random variables
contains exactly \eqn{k} peaks, with peaks also as defined by
\insertCite{Goldfeld65;textual}{skedastic}. The function is used in
\code{\link{ppeakdist}} to compute \eqn{p}-values for the Goldfeld-Quandt
nonparametric test for heteroskedasticity in a linear model.
}
\examples{
dpeakdist(5, 0:4)
sum(dpeakdist(5, 0:4))

}
\references{
{\insertAllCited{}}
}
\seealso{
\code{\link{ppeakdist}}, \code{\link{goldfeld_quandt}}
}
