% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbyla.R
\name{verbyla}
\alias{verbyla}
\title{Verbyla's Test for Heteroskedasticity in a Linear Regression Model}
\usage{
verbyla(mainlm, auxdesign = NULL)
}
\arguments{
\item{mainlm}{Either an object of \code{\link[base]{class}} "lm", or a list
of two components: a response vector and a design matrix (in that order).
If the latter, the design matrix must begin with a column of 1s if an
intercept is to be included in the linear model. Passing an object of
class "lm" is recommended in applications; passing a list containing
the data is recommended where optimising computational performance is
important.}

\item{auxdesign}{A \code{\link[base]{data.frame}} or
\code{\link[base]{matrix}} representing an auxiliary design matrix of
containing exogenous variables that (under alternative hypothesis) are
related to error variance, or a character "fitted.values" indicating
that the fitted \eqn{\hat{y}_i} values from OLS should be used.
If set to \code{NULL} (the default), the
design matrix of the original regression model is used. An intercept
is included in the auxiliary regression even if the first column of
\code{auxdesign} is not a vector of ones.}
}
\value{
An object of \code{\link[base]{class}} "htest". If object is not
   assigned, its attributes are displayed in the console as a
   \code{\link[tibble]{tibble}} using \code{\link[broom]{tidy}}.
}
\description{
This function implements the residual maximum likelihood test of
   \insertCite{Verbyla93;textual}{skedastic} for testing for heteroskedasticity
   in a linear regression model.
}
\details{
Verbyla's Test entails fitting a generalised auxiliary regression model
in which the response variable is the vector of standardised squared
residuals \eqn{e_i^2/\hat{\sigma}^2} from the original OLS
model and the design matrix is some function of \eqn{Z}, an
\eqn{n \times q} matrix consisting of \eqn{q} exogenous variables, appended
to a column of ones.
The test statistic is half the residual sum of squares from this generalised
auxiliary regression. Under the null hypothesis of homoskedasticity, the
distribution of the test statistic is asymptotically chi-squared with
\eqn{q} degrees of freedom. The test is right-tailed.
}
\examples{
mtcars_lm <- lm(mpg ~ wt + qsec + am, data = mtcars)
verbyla(mtcars_lm)
verbyla(mtcars_lm, auxdesign = "fitted.values")

}
\references{
{\insertAllCited{}}
}
