% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_lm.R
\name{white_lm}
\alias{white_lm}
\title{White's Test for Heteroskedasticity in a Linear Regression Model}
\usage{
white_lm(mainlm, interactions = FALSE)
}
\arguments{
\item{mainlm}{Either an object of \code{\link[base]{class}} "lm", or a list
of two components: a response vector and a design matrix (in that order).
If the latter, the design matrix must begin with a column of 1s if an
intercept is to be included in the linear model. Passing an object of
class "lm" is recommended in applications; passing a list containing
the data is recommended where optimising computational performance is
important.}

\item{interactions}{A logical. Should two-way interactions between explanatory
variables be included in the auxiliary regression? Defaults to
\code{FALSE}, since when interaction terms are present the test is not a
pure test of heteroskedasticity but also of model specification.}
}
\value{
An object of \code{\link[base]{class}} "htest". If object is not
   assigned, its attributes are displayed in the console as a
   \code{\link[tibble]{tibble}} using \code{\link[broom]{tidy}}.
}
\description{
This function implements the popular method of
   \insertCite{White80;textual}{skedastic} for testing for heteroskedasticity
   in a linear regression model.
}
\details{
White's Test entails fitting an auxiliary regression model in which the
response variable is the vector of squared residuals from the original
model and the design matrix includes the original explanatory variables,
their squares, and (optionally) their two-way interactions. The test
statistic is the number of observations multiplied by the coefficient of
determination from the auxiliary regression model:
\deqn{T = n r_{\mathrm{aux}}^2}
White's Test is thus a special case of the method of
\insertCite{Breusch79;textual}{skedastic}. Under the null hypothesis of
   homoskedasticity, the distribution of the test statistic is
   asymptotically chi-squared with \code{parameter} degrees of freedom.
   The test is right-tailed.
}
\examples{
n <- 20
p <- 4
set.seed(9586)
X <- matrix(data = runif(n * (p - 1)), nrow = n, ncol = p - 1)
# Response values generated under homoskedasticity
y_H0 <- rnorm(n, mean = 1 + rowSums(X), sd = 1)
white_lm(lm(y_H0 ~ X))
white_lm(lm(y_H0 ~ X), interactions = T)
# Response values generated under heteroskedasticity associated with X
y_HA <- rnorm(n, mean = 1 + rowSums(X), sd = rowSums(X ^ 2))
white_lm(lm(y_HA ~ X))
white_lm(lm(y_HA ~ X), interactions = T)

}
\references{
{\insertAllCited{}}
}
\seealso{
This function should not be confused with
  \code{\link[het.test:whites.htest]{het.test::whites.htest}} and
  \code{\link[tseries:white.test]{tseries::white.test}}, which do \emph{not}
  implement the method of \insertCite{White80;textual}{skedastic} for
  testing for heteroskedasticity in a linear model.
}
