% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{countpeaks}
\alias{countpeaks}
\title{Count peaks in a data series}
\usage{
countpeaks(x)
}
\arguments{
\item{x}{A double vector.}
}
\value{
An integer value between \code{0} and \code{length(x) - 1}
   representing the number of peaks in the series.
}
\description{
This function computes the number of peaks in a univariate data series, with
peak defined as per \insertCite{Goldfeld65;textual}{skedastic}. The function
is used in the Goldfeld-Quandt nonparametric test for heteroskedasticity in
a linear model.
}
\examples{
set.seed(9586)
countpeaks(rnorm(20))

}
\references{
{\insertAllCited{}}
}
\seealso{
\code{\link{goldfeld_quandt}}
}
